backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
qrcode
python-multipart
cloudinary
python-dotenv
pydantic
aiofiles
